+++ frontend/src/APIs/APIs.js
@@ -1,29 +1,8 @@
 import { requestOptions } from "./configAPI";
 
-export function GET_AVATAR(type, id) {
-  return requestOptions({
-    url: `/profile/${type}/${id}/avatar`,
-    isAuthenticated: false,
-  });
-}
-
-export function GET_PROFILE(type, id) {
-  return requestOptions({
-    url: `/profile/${type}/${id}`,
-  });
-}
-
-export function PATCH_PROFILE(type, id, body) {
-  return requestOptions({
-    url: `/profile/${type}/${id}`,
-    method: "PATCH",
-    body,
-  });
-}
-
 export function GET_OPPORTUNITY(id) {
   return requestOptions({
-    url: `/opportunity/${id}`,
+    url: `/opportunity/${id}`
   });
 }
 
@@ -34,7 +13,7 @@ export function GET_LIST_OPPORTUNITY(page = 0, name = "") {
   }
 
   return requestOptions({
-    url: uri,
+    url: uri
   });
 }
 
@@ -43,4 +22,4 @@ export function POST_APPLY(id) {
     url: `/opportunity/${id}/apply`,
     method: "POST"
   });
-};
+}
