+++ frontend/src/components/Profile/CompanyProfile/CompanyProfile.jsx
@@ -1,9 +1,9 @@
-import React, { useState } from 'react';
-import { useEffect } from 'react';
-import ProfileInfo from '../ProfileInfo';
-import useFetch from '../../../Hooks/useFetch';
-import { PATCH_PROFILE } from '../../../APIs/APIs';
-import { linkFormatter } from '../../../utils/formatters';
+import React, { useState } from "react";
+import { useEffect } from "react";
+import ProfileInfo from "../ProfileInfo";
+import useFetch from "../../../Hooks/useFetch";
+import { PATCH_PROFILE } from "../../../APIs/profileAPI";
+import { linkFormatter } from "../../../utils/formatters";
 
 const CompanyProfile = ({ data, canEdit, profileId }) => {
   const { request } = useFetch();
@@ -18,19 +18,19 @@ const CompanyProfile = ({ data, canEdit, profileId }) => {
     setInfo({
       about: { text: "Sobre", value: data.about, type: "textarea" },
       location: { text: "Localização", value: data.location },
-      category: {text: "Categoria", value: data.category, required: true},
+      category: { text: "Categoria", value: data.category, required: true },
       site: {
         text: "Site",
         value: data.site,
         formatter: linkFormatter,
-        placeholder: "https://sitedaempresa.com.br",
+        placeholder: "https://sitedaempresa.com.br"
       },
       linkedin: {
         text: "LinkedIn",
         value: data.linkedin,
         formatter: linkFormatter,
-        placeholder: "https://linkedin.com/company/empresa",
-      },
+        placeholder: "https://linkedin.com/company/empresa"
+      }
     });
   }, [data]);
