+++ frontend/src/components/Store/Provider.jsx
@@ -4,7 +4,7 @@ import Context from "./Context";
 import useStorage from "../../utils/useStorage";
 import jwt_decode from "jwt-decode";
 import logo from "../../assets/images/Logo2RecruIT.svg";
-import { GET_AVATAR } from "../../APIs/APIs";
+import { GET_AVATAR } from "../../APIs/profileAPI";
 import useFetch from "../../Hooks/useFetch";
 
 const StoreProvider = ({ children }) => {
@@ -37,7 +37,7 @@ const StoreProvider = ({ children }) => {
         apptoken,
         setToken,
         user,
-        avatar,
+        avatar
       }}
     >
       {children}
